adk_backend/requirements.txt
google-adk
requests
litellm